Move proxy_headers behavior into middleware
