+++ tests/protocols/test_http.py
@@ -1,50 +1,10 @@
+from tests.response import Response
 from uvicorn.protocols.http.h11_impl import H11Protocol
 from uvicorn.protocols.http.httptools_impl import HttpToolsProtocol
 from uvicorn.protocols.websockets.wsproto_impl import WSProtocol
 import asyncio
 import h11
 import pytest
-
-
-class Response:
-    charset = "utf-8"
-
-    def __init__(self, content, status_code=200, headers=None, media_type=None):
-        self.body = self.render(content)
-        self.status_code = status_code
-        self.headers = headers or {}
-        self.media_type = media_type
-        self.set_content_type()
-        self.set_content_length()
-
-    async def __call__(self, receive, send) -> None:
-        await send(
-            {
-                "type": "http.response.start",
-                "status": self.status_code,
-                "headers": [
-                    [key.encode(), value.encode()]
-                    for key, value in self.headers.items()
-                ],
-            }
-        )
-        await send({"type": "http.response.body", "body": self.body})
-
-    def render(self, content) -> bytes:
-        if isinstance(content, bytes):
-            return content
-        return content.encode(self.charset)
-
-    def set_content_length(self):
-        if "content-length" not in self.headers:
-            self.headers["content-length"] = str(len(self.body))
-
-    def set_content_type(self):
-        if self.media_type is not None and "content-type" not in self.headers:
-            content_type = self.media_type
-            if content_type.startswith("text/") and self.charset is not None:
-                content_type += "; charset=%s" % self.charset
-            self.headers["content-type"] = content_type
 
 
 SIMPLE_GET_REQUEST = b"\r\n".join([b"GET / HTTP/1.1", b"Host: example.org", b"", b""])
@@ -580,32 +540,6 @@ def test_root_path(protocol_cls):
     protocol.loop.run_one()
     assert b"HTTP/1.1 200 OK" in protocol.transport.buffer
     assert b"Path: /app/" in protocol.transport.buffer
-
-
-@pytest.mark.parametrize("protocol_cls", [HttpToolsProtocol, H11Protocol])
-def test_proxy_headers(protocol_cls):
-    def app(scope):
-        scheme = scope["scheme"]
-        host, port = scope["client"]
-        addr = "%s://%s:%d" % (scheme, host, port)
-        return Response("Remote: " + addr, media_type="text/plain")
-
-    REQUEST = b"\r\n".join(
-        [
-            b"GET / HTTP/1.1",
-            b"Host: example.org",
-            b"X-Forwarded-Proto: https",
-            b"X-Forwarded-For: 1.2.3.4",
-            b"X-Forwarded-Port: 567",
-            b"",
-            b"",
-        ]
-    )
-    protocol = get_connected_protocol(app, protocol_cls, proxy_headers=True)
-    protocol.data_received(REQUEST)
-    protocol.loop.run_one()
-    assert b"HTTP/1.1 200 OK" in protocol.transport.buffer
-    assert b"Remote: https://1.2.3.4:567" in protocol.transport.buffer
 
 
 @pytest.mark.parametrize("protocol_cls", [HttpToolsProtocol, H11Protocol])
