+++ uvicorn/protocols/http/httptools_impl.py
@@ -112,7 +112,6 @@ class HttpToolsProtocol(asyncio.Protocol):
         logger=None,
         access_log=True,
         ws_protocol_class=None,
-        proxy_headers=False,
         root_path="",
         limit_concurrency=None,
         timeout_keep_alive=5,
@@ -126,7 +125,6 @@ class HttpToolsProtocol(asyncio.Protocol):
         self.access_log = access_log and (self.logger.level <= logging.INFO)
         self.parser = httptools.HttpRequestParser(self)
         self.ws_protocol_class = ws_protocol_class
-        self.proxy_headers = proxy_headers
         self.root_path = root_path
         self.limit_concurrency = limit_concurrency
 
@@ -263,10 +261,6 @@ class HttpToolsProtocol(asyncio.Protocol):
         http_version = self.parser.get_http_version()
         if http_version != "1.1":
             self.scope["http_version"] = http_version
-        if self.proxy_headers:
-            scheme, client = _get_remote_from_proxy(self.scope)
-            self.scope["scheme"] = scheme
-            self.scope["client"] = client
         if self.parser.should_upgrade():
             return
